+++ lib/src/widgets/size_listener.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:nb_utils/nb_utils.dart';
 
 /// [SizeListener] Listen to your child widget's size
 class SizeListener extends StatefulWidget {
@@ -24,37 +23,22 @@ class _SizeListenerState extends State<SizeListener> {
   GlobalKey widgetKey = GlobalKey();
   Size? oldSize;
 
-  @override
-  void initState() {
-    afterBuildCreated(() => postFrameCallback);
-    super.initState();
-  }
-
   void postFrameCallback(_) async {
-    var context = widgetKey.currentContext!;
+    BuildContext? context = widgetKey.currentContext;
+    if (context == null) return;
 
-    await Future.delayed(widget.delayDuration ?? Duration(milliseconds: 200));
-    Size newSize = context.size!;
+    await Future.delayed(widget.delayDuration ?? Duration(milliseconds: 0));
 
-    if (newSize == Size.zero) return;
-    if (oldSize == newSize) return;
+    var newSize = context.size;
+    if (oldSize == newSize || newSize == null) return;
 
     oldSize = newSize;
     widget.onSizeChange.call(newSize);
   }
 
   @override
-  void didUpdateWidget(covariant SizeListener oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    afterBuildCreated(() => postFrameCallback);
-  }
-
-  @override
   Widget build(BuildContext context) {
     SchedulerBinding.instance.addPostFrameCallback(postFrameCallback);
-    return Container(
-      key: widgetKey,
-      child: widget.child,
-    );
+    return Container(key: widgetKey, child: widget.child);
   }
 }